updated architecture + twig template applied
