--- index.php
@@ -1,8 +1,0 @@
-<?php
-
-require_once __DIR__ . '/vendor/autoload.php';
-
-use App\Controller\Router;
-
-$router = new Router();
-$router->routerRequest();
