+++ public/index.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+require_once '../vendor/autoload.php';
+
+use App\Service\Router;
+
+$whoops = new \Whoops\Run;
+$whoops->pushHandler(new \Whoops\Handler\PrettyPageHandler);
+$whoops->register();
+
+$router = new Router();
+$router->routerRequest();
