--- public/view/view.php
@@ -1,12 +1,0 @@
-<?php
-namespace Pub\View;
-
-class View
-{
-
-    // Génère et affiche la vue
-    public function generate()
-    {
-        require_once 'layout.php';
-    }
-}
