+++ src/View/View.php
@@ -1,0 +1,23 @@
+<?php
+declare(strict_types=1);
+
+namespace App\View;
+
+use Twig\Environment;
+use Twig\Loader\FilesystemLoader;
+
+class View
+{
+    private $twig;
+
+    public function __construct()
+    {
+        $loader = new FilesystemLoader('../templates');
+        $this->twig = new Environment($loader);
+    }
+
+    public function render(): void
+    {
+        echo $this->twig->render('frontoffice/layout.twig');
+    }
+}
