+++ templates/frontoffice/layout.twig
@@ -2,7 +2,6 @@
 <html lang="fr">
 
 <head>
-
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
   <meta name="description" content="">
@@ -11,16 +10,15 @@
   <title>Mon blog de dev, basé sur bootstrap</title>
 
   <!-- Bootstrap core CSS -->
-  <link href="public/css/bootstrap.min.css" rel="stylesheet">
+  <link href="css/bootstrap.min.css" rel="stylesheet">
 
   <!-- Custom fonts for this template -->
-  <link href="public/css/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
+  <link href="css/fontawesome-free/css/all.min.css" rel="stylesheet" type="text/css">
   <link href='https://fonts.googleapis.com/css?family=Lora:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
   <link href='https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
 
   <!-- Custom styles for this template -->
-  <link href="public/css/clean-blog.min.css" rel="stylesheet">
-
+  <link href="css/clean-blog.min.css" rel="stylesheet">
 </head>
 
 <body>
@@ -39,13 +37,13 @@
             <a class="nav-link" href="index.php">Accueil</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="about.html">A Propos</a>
+            <a class="nav-link" href="index.php">A Propos</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="post.html">Post</a>
+            <a class="nav-link" href="index.php">Post</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="contact.html">Contact</a>
+            <a class="nav-link" href="index.php">Contact</a>
           </li>
         </ul>
       </div>
@@ -53,7 +51,7 @@
   </nav>
 
   <!-- Page Header -->
-  <header class="masthead" style="background-image: url('public/img/home-bg.jpg')">
+  <header class="masthead" style="background-image: url('img/home-bg.jpg')">
     <div class="overlay"></div>
     <div class="container">
       <div class="row">
@@ -173,12 +171,10 @@
   </footer>
 
   <!-- Bootstrap core JavaScript -->
-  <script src="public/js/jquery.min.js"></script>
-  <script src="public/js/bootstrap.bundle.min.js"></script>
+  <script src="js/jquery.min.js"></script>
+  <script src="js/bootstrap.bundle.min.js"></script>
 
   <!-- Custom scripts for this template -->
-  <script src="public/js/clean-blog.min.js"></script>
-
+  <script src="js/clean-blog.min.js"></script>
 </body>
-
 </html>
